Add initial reset to support repeat running NuCWriter
At the start of NuCWriter execution, an unconditional USB reset is
performed on the WinUSB device (supposed to be NUC980). Regardless
of whether NUC980 is in IBR execution state or previously executed
NuCWriter and is in XUSB state, the USB reset can bring NUC980
back to IBR, waiting for USB connection state."
This modification is intended to make NuCWriter suitable for
consecutive multiple executions.

Change-Id: I6ec82952b93c9312b04f29b85a8a271ad9a157c7