Fixed the issue with SPI NAND erase all.
In the command handshake for erasing SPI NAND, the xusb firmware responds
differently to 'partial erase' and 'erase all'. In the case of partial erase,
the xusb ack responds with the erase progress percentage, whereas in the
case of erase all, the xusb ack responds with the current erase block number.
These two scenarios require different handling.
This patch addresses the handling of erase all scenarios.

Change-Id: Ib9e8081131b59d9febce45bfe501faa746b513ec